Create e.cpp
A permutation is a sequence of n integers from 1 to n, in which all numbers occur exactly once. For example, [1], [3,5,2,1,4], [1,3,2] are permutations, and [2,3,2], [4,3,1], [0] are not.

Polycarp was presented with a permutation p of numbers from 1 to n. However, when Polycarp came home, he noticed that in his pocket, the permutation p had turned into an array q according to the following rule:

qi=max(p1,p2,…,pi).
Now Polycarp wondered what lexicographically minimal and lexicographically maximal permutations could be presented to him.

An array a of length n is lexicographically smaller than an array b of length n if there is an index i (1≤i≤n) such that the first i−1 elements of arrays a and b are the same, and the i-th element of the array a is less than the i-th element of the array b. For example, the array a=[1,3,2,3] is lexicographically smaller than the array b=[1,3,4,2].

For example, if n=7 and p=[3,2,4,1,7,5,6], then q=[3,3,4,4,7,7,7] and the following permutations could have been as p initially:

[3,1,4,2,7,5,6] (lexicographically minimal permutation);
[3,1,4,2,7,6,5];
[3,2,4,1,7,5,6];
[3,2,4,1,7,6,5] (lexicographically maximum permutation).
For a given array q, find the lexicographically minimal and lexicographically maximal permutations that could have been originally presented to Polycarp.