Create icpc firecracker basic code writing.cpp
There are m firecrackers in the hooligan's pocket, the i-th firecracker explodes in si seconds after being lit.

The following events happen each second (sequentially, exactly in the following order):

firstly, the hooligan either moves into an adjacent cell (from the cell i, he can move to the cell (i+1) or to the cell (i−1), and he cannot leave the corridor) or stays in the cell he is currently. If the hooligan doesn't move, he can light one of his firecrackers and drop it. The hooligan can't move into the cell where the guard is;
secondly, some firecrackers that were already dropped may explode. Formally, if the firecracker j is dropped on the T-th second, then it will explode on the (T+sj)-th second (for example, if a firecracker with sj=2 is dropped on the 4-th second, it explodes on the 6-th second);
finally, the guard moves one cell closer to the hooligan. If the guard moves to the cell where the hooligan is, the hooligan is caught.
Obviously, the hooligan will be caught sooner or later, since the corridor is finite. His goal is to see the maximum number of firecrackers explode before he is caught; that is, he will act in order to maximize the number of firecrackers that explodes before he is caught.

Your task is to calculate the number of such firecrackers, if the hooligan acts optimally.